requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
passlib==1.7.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
Pyrebase4==4.4.0
python-jwt==2.0.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
